Web/ManyFiles/app/requirements.txt
aniso8601==4.0.1
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Limiter==1.0.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10
limits==1.3
MarkupSafe==1.1.0
pytz==2018.7
six==1.12.0
Werkzeug==0.14.1
yara-python==3.8.1
flask-restful-swagger